Renamed base class to use a base prefix rather than a suffix Moved ensure argument to the helpers namespace Moved the db set extensions to an internal namespace
