+++ src/EntityFrameworkCore 3.0.0/EntityFrameworkCore.Testing.Common.Tests/BaseForDbContextTests.cs
@@ -11,7 +11,7 @@ using NUnit.Framework;
 namespace EntityFrameworkCore.Testing.Common.Tests
 {
     [TestFixture]
-    public abstract class DbContextTestsBase<TDbContext> : TestBase
+    public abstract class BaseForDbContextTests<TDbContext> : BaseForTests
         where TDbContext : DbContext
     {
         protected TDbContext MockedDbContext;
