+++ src/EntityFrameworkCore 3.0.0/EntityFrameworkCore.Testing.Common.Tests/BaseForMockedDbSetQueryProviderTests.cs
@@ -11,8 +11,8 @@ using NUnit.Framework;
 namespace EntityFrameworkCore.Testing.Common.Tests
 {
     [TestFixture]
-    public abstract class MockedDbSetQueryProviderTestsBase<TEntity> : QueryableTestsBase<TEntity>
-        where TEntity : TestEntityBase
+    public abstract class BaseForMockedDbSetQueryProviderTests<TEntity> : BaseForQueryableTests<TEntity>
+        where TEntity : BaseTestEntity
     {
         protected abstract void AddFromSqlRawResult(DbSet<TEntity> mockedDbSet, IEnumerable<TEntity> expectedResult);
         protected abstract void AddFromSqlRawResult(DbSet<TEntity> mockedDbSet, string sql, IEnumerable<TEntity> expectedResult);
