+++ src/EntityFrameworkCore 3.0.0/EntityFrameworkCore.Testing.Common/Helpers/EnsureArgument.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace EntityFrameworkCore.Testing.Common
+namespace EntityFrameworkCore.Testing.Common.Helpers
 {
     /// <summary>A helper to perform checks on arguments.</summary>
     public class EnsureArgument
