+++ src/EntityFrameworkCore 3.0.0/EntityFrameworkCore.Testing.Moq.Tests/BaseForDbSetTests.cs
@@ -8,7 +8,7 @@ using NUnit.Framework;
 namespace EntityFrameworkCore.Testing.Moq.Tests
 {
     [TestFixture]
-    public abstract class DbSetTestsBase<T> : DbSetTestsBase<TestDbContext, T> where T : TestEntityBase
+    public abstract class BaseForDbSetTests<T> : BaseForDbSetTests<TestDbContext, T> where T : BaseTestEntity
     {
         protected override TestDbContext CreateMockedDbContext()
         {
