+++ src/EntityFrameworkCore 3.0.0/EntityFrameworkCore.Testing.NSubstitute/Extensions/DbSetExtensions.Internal.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
 using System.Threading;
-using EntityFrameworkCore.Testing.Common;
+using EntityFrameworkCore.Testing.Common.Helpers;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using NSubstitute;
@@ -12,13 +12,11 @@ using NSubstitute.Extensions;
 
 namespace EntityFrameworkCore.Testing.NSubstitute.Extensions
 {
-    /// <summary>Extensions for the db set type.</summary>
+    /// <summary>
+    ///     Extensions for db sets.
+    /// </summary>
     public static class DbSetExtensions
     {
-        /// <summary>Creates and sets up a mocked db set.</summary>
-        /// <typeparam name="TEntity">The entity type.</typeparam>
-        /// <param name="dbSet">The db set to mock/proxy.</param>
-        /// <returns>A mocked db set.</returns>
         internal static DbSet<TEntity> CreateMockedDbSet<TEntity>(this DbSet<TEntity> dbSet)
             where TEntity : class
         {
